Implement insert and insert tail functions
For insert, first check if the head exists.
Then, allocate memory for an element_t node.
And then, use strdup to allocate memory for the string value.
Check if the allocation is successful.
Use the list_add macro to add the node at the front of the queue.
For insert tail, reuse insert to implement.

Change-Id: I6fe1fd9b524f56ea921bdc83fed142ef1c57dcfb